requirements.txt
# Core ML/DL Libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
timm>=0.9.0

# Segmentation & Vision
segmentation-models-pytorch>=0.3.3
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# Data Handling
numpy>=1.24.0
pandas>=2.0.0
roboflow>=1.1.0

# Evaluation Metrics
scikit-image>=0.21.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
tensorboard>=2.13.0

# Optional: For PDF reading (if needed)
PyPDF2>=3.0.0
pdfplumber>=0.10.0

